+++ dcli/lib/src/script/my_yaml.dart
@@ -6,7 +6,6 @@
 
 // ignore_for_file: deprecated_member_use
 
-import 'dart:cli';
 import 'dart:io';
 
 import 'package:yaml/yaml.dart' as y;
@@ -22,7 +21,7 @@ class MyYaml {
   /// reads yaml from file.
   MyYaml.fromFile(String path) {
     // ignore: discarded_futures
-    final contents = waitFor<String>(File(path).readAsString());
+    final contents = File(path).readAsStringSync();
     _document = _load(contents);
   }