+++ dcli_core/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.0.1-alpha.9
+- added method withEnvironment - this may be problematic as it easy to 
+use with an async callback which will end in tears.
+
 # 4.0.1-alpha.8
 - fixed the sdk range for dcli_common
 - switch to activating dcli_sdk instead of dcli.